Refactoring skeleton
