--- collection/document_exists_test.go
@@ -1,80 +1,0 @@
-package collection_test
-
-import (
-	"encoding/json"
-	"fmt"
-	"github.com/kuzzleio/sdk-go/collection"
-
-	"github.com/kuzzleio/sdk-go/internal"
-	"github.com/kuzzleio/sdk-go/kuzzle"
-	"github.com/kuzzleio/sdk-go/types"
-	"github.com/stretchr/testify/assert"
-	"testing"
-)
-
-func TestDocumentExistsEmptyId(t *testing.T) {
-	c := &internal.MockedConnection{
-		MockSend: func(query []byte, options types.QueryOptions) *types.KuzzleResponse {
-			return &types.KuzzleResponse{Error: &types.KuzzleError{Message: "Collection.DocumentExists: document id required"}}
-		},
-	}
-	k, _ := kuzzle.NewKuzzle(c, nil)
-
-	_, err := collection.NewCollection(k, "collection", "index").DocumentExists("", nil)
-	assert.NotNil(t, err)
-}
-
-func TestDocumentExistsError(t *testing.T) {
-	c := &internal.MockedConnection{
-		MockSend: func(query []byte, options types.QueryOptions) *types.KuzzleResponse {
-			return &types.KuzzleResponse{Error: &types.KuzzleError{Message: "Unit test error"}}
-		},
-	}
-	k, _ := kuzzle.NewKuzzle(c, nil)
-
-	_, err := collection.NewCollection(k, "collection", "index").DocumentExists("myId", nil)
-	assert.NotNil(t, err)
-}
-
-func TestDocumentExists(t *testing.T) {
-	type Document struct {
-		Title string
-	}
-
-	id := "myId"
-
-	c := &internal.MockedConnection{
-		MockSend: func(query []byte, options types.QueryOptions) *types.KuzzleResponse {
-			parsedQuery := &types.KuzzleRequest{}
-			json.Unmarshal(query, parsedQuery)
-
-			assert.Equal(t, "document", parsedQuery.Controller)
-			assert.Equal(t, "exists", parsedQuery.Action)
-			assert.Equal(t, "index", parsedQuery.Index)
-			assert.Equal(t, "collection", parsedQuery.Collection)
-			assert.Equal(t, id, parsedQuery.Id)
-
-			r, _ := json.Marshal(true)
-			return &types.KuzzleResponse{Result: r}
-		},
-	}
-	k, _ := kuzzle.NewKuzzle(c, nil)
-
-	res, _ := collection.NewCollection(k, "collection", "index").DocumentExists(id, nil)
-	assert.Equal(t, true, res)
-}
-
-func ExampleCollection_DocumentExists() {
-	id := "myId"
-	c := &internal.MockedConnection{}
-	k, _ := kuzzle.NewKuzzle(c, nil)
-
-	res, err := collection.NewCollection(k, "collection", "index").DocumentExists(id, nil)
-
-	if err != nil {
-		fmt.Println(err.Error())
-		return
-	}
-
-	fmt.Println(res)
-}
