+++ document/document.go
@@ -1,4 +1,4 @@
-package collection
+package document
 
 import (
 	"encoding/json"
